chore: separate functions
