+++ packages/shared/provideLocal/index.ts
@@ -1,0 +1,25 @@
+import { InjectionKey, getCurrentInstance, inject, provide } from 'vue-demi'
+import { localProvidedStateMap } from './map'
+
+/**
+ * On the basis of `provide`, it is allowed to directly call inject to obtain the value after call provide in the same component.
+ *
+ * @example
+ * ```ts
+ * provideLocal('MyInjectionKey', 1)
+ * const injectedValue = injectLocal('MyInjectionKey') // injectedValue === 1
+ * ```
+ */
+export const provideLocal: typeof provide = (key, value) => {
+  const instance = getCurrentInstance()
+  if (instance == null)
+    throw new Error('provideLocal must be called in setup')
+
+  if (!localProvidedStateMap.has(instance))
+    localProvidedStateMap.set(instance, Object.create(null))
+
+  const localProvidedState = localProvidedStateMap.get(instance)!
+  // @ts-expect-error allow InjectionKey as key
+  localProvidedState[key] = value
+  provide(key, value)
+}
