+++ packages/shared/provideLocal/map.ts
@@ -1,0 +1,3 @@
+import { getCurrentInstance } from 'vue-demi';
+
+export const localProvidedStateMap = new WeakMap<NonNullable<ReturnType<typeof getCurrentInstance>>, Record<string | symbol, any>>();
